Represent validated source archives before persistence
