+++ crates/uv-distribution/src/source/mod.rs
@@ -15,9 +15,8 @@ use std::str::FromStr;
 use std::sync::Arc;
 
 use fs_err::tokio as fs;
-use futures::{FutureExt, TryStreamExt};
+use futures::FutureExt;
 use reqwest::{Response, StatusCode};
-use tokio_util::compat::FuturesAsyncReadCompatExt;
 use tracing::{Instrument, debug, info_span, instrument, warn};
 use url::Url;
 
@@ -35,7 +34,6 @@ use uv_distribution_types::{
     ExtraBuildRequirement, GitDirectorySourceUrl, GitPathSourceUrl, HashPolicy, Hashed, IndexUrl,
     PathSourceUrl, RemoteSource, RequirementSource, RequiresPython, SourceDist, SourceUrl,
 };
-use uv_extract::hash::Hasher;
 use uv_fs::{Simplified, rename_with_retry, write_atomic};
 use uv_git::{Fetch, GIT_LFS, GitError, GitHttpSettings, GitResolver};
 use uv_git_types::{GitHubRepository, GitOid, GitUrl};
@@ -54,10 +52,12 @@ use crate::hash::http_hash_algorithms;
 use crate::metadata::{ArchiveMetadata, GitWorkspaceMember, Metadata};
 use crate::source::built_wheel_metadata::{BuiltWheelFile, BuiltWheelMetadata};
 use crate::source::revision::Revision;
+use crate::source::validated_archive::ValidatedSourceArchive;
 use crate::{Reporter, RequiresDist};
 
 mod built_wheel_metadata;
 mod revision;
+mod validated_archive;
 
 /// Access distribution metadata without requiring a build interpreter.
 ///
@@ -2813,7 +2813,7 @@ impl<'a, T: BuildContext> SourceDistributionBuilder<'a, T> {
             .await
     }
 
-    /// Download and unzip a source distribution into the cache from an HTTP response.
+    /// Download, extract, validate, and persist a source distribution into the cache.
     async fn download_archive(
         &self,
         response: Response,
@@ -2822,87 +2822,19 @@ impl<'a, T: BuildContext> SourceDistributionBuilder<'a, T> {
         target: &Path,
         algorithms: &[HashAlgorithm],
     ) -> Result<(Vec<HashDigest>, u64), Error> {
-        let temp_dir = tempfile::tempdir_in(
-            self.build_context
-                .cache()
-                .bucket(CacheBucket::SourceDistributions),
+        let archive = ValidatedSourceArchive::extract_http(
+            response,
+            source,
+            ext,
+            self.build_context.cache(),
+            algorithms,
         )
-        .map_err(Error::CacheWrite)?;
-
-        let reader = response
-            .bytes_stream()
-            .map_err(std::io::Error::other)
-            .into_async_read();
-
-        // Create a hasher for each hash algorithm.
-        let mut hashers = algorithms
-            .iter()
-            .copied()
-            .map(Hasher::from)
-            .collect::<Vec<_>>();
-        let mut hasher = uv_extract::hash::HashReader::new(reader.compat(), &mut hashers);
-
-        // Download and unzip the source distribution into a temporary directory.
-        let span = info_span!("download_source_dist", source_dist = %source);
-        uv_extract::stream::archive(&mut hasher, ext, temp_dir.path())
-            .await
-            .map_err(|err| Error::Extract(source.to_string(), err))?;
-        drop(span);
-
-        let expected_size = match source {
-            BuildableSource::Dist(SourceDist::Registry(dist)) if dist.size_is_authoritative => {
-                dist.size()
-            }
-            BuildableSource::Dist(SourceDist::DirectUrl(dist)) => dist.size(),
-            _ => None,
-        };
-
-        // If necessary, exhaust the reader to compute the hash or validate the archive size.
-        if !algorithms.is_empty() || expected_size.is_some() {
-            hasher.finish().await.map_err(Error::HashExhaustion)?;
-        }
-        if let Some(expected) = expected_size
-            && hasher.bytes_read() != expected
-        {
-            return Err(Error::MismatchedSize {
-                distribution: source.to_string(),
-                expected,
-                actual: hasher.bytes_read(),
-            });
-        }
-
-        let size = hasher.bytes_read();
-        let hashes = hashers.into_iter().map(HashDigest::from).collect();
-
-        // Extract the top-level directory.
-        let extracted = match uv_extract::strip_component(temp_dir.path()) {
-            Ok(top_level) => top_level,
-            Err(uv_extract::Error::NonSingularArchive(_)) => temp_dir.keep(),
-            Err(err) => {
-                return Err(Error::Extract(
-                    temp_dir.path().to_string_lossy().into_owned(),
-                    err,
-                ));
-            }
-        };
-
-        // Persist it to the cache.
-        fs_err::tokio::create_dir_all(target.parent().expect("Cache entry to have parent"))
-            .await
-            .map_err(Error::CacheWrite)?;
-        if let Err(err) = rename_with_retry(extracted, target).await {
-            // If the directory already exists, accept it.
-            if err.kind() == std::io::ErrorKind::AlreadyExists {
-                warn!("Directory already exists: {}", target.display());
-            } else {
-                return Err(Error::CacheWrite(err));
-            }
-        }
-
-        Ok((hashes, size))
+        .await?;
+        let metadata = archive.persist(target).await?;
+        Ok((metadata.hashes, metadata.size))
     }
 
-    /// Extract a local archive, and store it at the given [`CacheEntry`].
+    /// Extract, validate, and persist a local source archive into the cache.
     async fn persist_archive(
         &self,
         path: &Path,
@@ -2910,64 +2842,14 @@ impl<'a, T: BuildContext> SourceDistributionBuilder<'a, T> {
         target: &Path,
         algorithms: &[HashAlgorithm],
     ) -> Result<Vec<HashDigest>, Error> {
-        debug!("Unpacking for build: {}", path.display());
-
-        let temp_dir = tempfile::tempdir_in(
-            self.build_context
-                .cache()
-                .bucket(CacheBucket::SourceDistributions),
+        let archive = ValidatedSourceArchive::extract_local(
+            path,
+            ext,
+            self.build_context.cache(),
+            algorithms,
         )
-        .map_err(Error::CacheWrite)?;
-        let reader = fs_err::tokio::File::open(&path)
-            .await
-            .map_err(Error::CacheRead)?;
-
-        // Create a hasher for each hash algorithm.
-        let mut hashers = algorithms
-            .iter()
-            .copied()
-            .map(Hasher::from)
-            .collect::<Vec<_>>();
-        let mut hasher = uv_extract::hash::HashReader::new(reader, &mut hashers);
-
-        // Unzip the archive into a temporary directory.
-        uv_extract::stream::archive(&mut hasher, ext, &temp_dir.path())
-            .await
-            .map_err(|err| Error::Extract(temp_dir.path().to_string_lossy().into_owned(), err))?;
-
-        // If necessary, exhaust the reader to compute the hash.
-        if !algorithms.is_empty() {
-            hasher.finish().await.map_err(Error::HashExhaustion)?;
-        }
-
-        let hashes = hashers.into_iter().map(HashDigest::from).collect();
-
-        // Extract the top-level directory from the archive.
-        let extracted = match uv_extract::strip_component(temp_dir.path()) {
-            Ok(top_level) => top_level,
-            Err(uv_extract::Error::NonSingularArchive(_)) => temp_dir.path().to_path_buf(),
-            Err(err) => {
-                return Err(Error::Extract(
-                    temp_dir.path().to_string_lossy().into_owned(),
-                    err,
-                ));
-            }
-        };
-
-        // Persist it to the cache.
-        fs_err::tokio::create_dir_all(target.parent().expect("Cache entry to have parent"))
-            .await
-            .map_err(Error::CacheWrite)?;
-        if let Err(err) = rename_with_retry(extracted, target).await {
-            // If the directory already exists, accept it.
-            if err.kind() == std::io::ErrorKind::DirectoryNotEmpty {
-                warn!("Directory already exists: {}", target.display());
-            } else {
-                return Err(Error::CacheWrite(err));
-            }
-        }
-
-        Ok(hashes)
+        .await?;
+        Ok(archive.persist(target).await?.hashes)
     }
 
     /// For Git directories, we check them out into the cache, so we need to avoid workspace
